requirements.txt
Django==1.10.3
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
factory-boy==2.7.0
fake-factory==0.7.2
mysqlclient==1.3.9
Pillow==3.4.2
python-dateutil==2.6.0
six==1.10.0
sqlparse==0.2.2
